Removed InferenceEngine dependency on Legacy library (#1932)
* Split libraries

* Small fixes

* Fixes after self-review

* Minimized legacy usage

* Fixed IE_CPU.onnx_prior_box

* Fixed linkage

* Temp WA for private plugins

* removed comments

* Reverted OneHot changes back
