+++ inference-engine/cmake/config.cmake.in
@@ -3,8 +3,7 @@
 #
 
 if(DEFINED IE_MAIN_SOURCE_DIR AND TARGET inference_engine)
-    set(InferenceEngine_LIBRARIES inference_engine_legacy inference_engine
-                                  inference_engine_c_api)
+    set(InferenceEngine_LIBRARIES inference_engine inference_engine_c_api)
 else()
     include("${CMAKE_CURRENT_LIST_DIR}/targets.cmake")
     if(NOT MSVC)
@@ -29,8 +28,7 @@ else()
     endif()
 
     get_target_property(InferenceEngine_INCLUDE_DIRS IE::inference_engine INTERFACE_INCLUDE_DIRECTORIES)
-    set(InferenceEngine_LIBRARIES IE::inference_engine_legacy IE::inference_engine
-                                  IE::inference_engine_c_api)
+    set(InferenceEngine_LIBRARIES IE::inference_engine IE::inference_engine_c_api)
 
     foreach(library IN LISTS InferenceEngine_LIBRARIES)
         if(CMAKE_CROSSCOMPILING AND NOT MSVC)
