+++ inference-engine/ie_bridges/python/CMakeLists.txt
@@ -18,10 +18,12 @@ elseif(ARCH STREQUAL "i386")
 endif()
 
 if(DEFINED IE_MAIN_SOURCE_DIR)
-    set(InferenceEngine_LIBRARIES inference_engine inference_engine_plugin_api)
+    set(InferenceEngine_LIBRARIES inference_engine)
+    list(APPEND InferenceEngine_LIBRARIES inference_engine_legacy)  # TODO: remove in PR #2755
 else()
     find_package(InferenceEngineDeveloperPackage REQUIRED)
-    set(InferenceEngine_LIBRARIES IE::inference_engine IE::inference_engine_plugin_api)
+    set(InferenceEngine_LIBRARIES IE::inference_engine)
+    list(APPEND InferenceEngine_LIBRARIES IE::inference_engine_legacy) # TODO: remove in PR #2755
 endif()
 
 if(UNIX)
