+++ inference-engine/src/inference_engine/file_utils.cpp
@@ -18,12 +18,46 @@
 # include <limits.h>
 # include <unistd.h>
 # include <dlfcn.h>
+# ifdef ENABLE_UNICODE_PATH_SUPPORT
+#  include <locale>
+#  include <codecvt>
+# endif
 #else
 # if defined(WINAPI_FAMILY) && !WINAPI_PARTITION_DESKTOP
 #  error "Only WINAPI_PARTITION_DESKTOP is supported, because of GetModuleHandleEx[A|W]"
 # endif
 # include <Windows.h>
 #endif
+
+#ifdef ENABLE_UNICODE_PATH_SUPPORT
+
+std::string FileUtils::wStringtoMBCSstringChar(const std::wstring& wstr) {
+#ifdef _WIN32
+    int size_needed = WideCharToMultiByte(CP_UTF8, 0, &wstr[0], (int)wstr.size(), NULL, 0, NULL, NULL);  // NOLINT
+    std::string strTo(size_needed, 0);
+    WideCharToMultiByte(CP_UTF8, 0, &wstr[0], (int)wstr.size(), &strTo[0], size_needed, NULL, NULL);  // NOLINT
+    return strTo;
+#else
+    std::wstring_convert<std::codecvt_utf8<wchar_t>> wstring_decoder;
+    return wstring_decoder.to_bytes(wstr);
+#endif
+}
+
+std::wstring FileUtils::multiByteCharToWString(const char* str) {
+#ifdef _WIN32
+    int strSize = static_cast<int>(std::strlen(str));
+    int size_needed = MultiByteToWideChar(CP_UTF8, 0, str, strSize, NULL, 0);
+    std::wstring wstrTo(size_needed, 0);
+    MultiByteToWideChar(CP_UTF8, 0, str, strSize, &wstrTo[0], size_needed);
+    return wstrTo;
+#else
+    std::wstring_convert<std::codecvt_utf8<wchar_t>> wstring_encoder;
+    std::wstring result = wstring_encoder.from_bytes(str);
+    return result;
+#endif
+}
+
+#endif  // ENABLE_UNICODE_PATH_SUPPORT
 
 long long FileUtils::fileSize(const char* charfilepath) {
 #if defined(ENABLE_UNICODE_PATH_SUPPORT) && defined(_WIN32)
