+++ inference-engine/src/inference_engine/generic_ie.cpp
@@ -14,8 +14,8 @@
 #include <vector>
 
 #include "blob_factory.hpp"
+#include <ie_ngraph_utils.hpp>
 #include "shape_infer/ie_ishape_infer_extension.hpp"
-#include <legacy/ie_ngraph_utils.hpp>
 #include "ngraph/util.hpp"
 #include "ngraph/graph_util.hpp"
 #include "ngraph/validation_util.hpp"
