+++ inference-engine/src/mkldnn_plugin/mkldnn_graph_dumper.cpp
@@ -5,7 +5,7 @@
 #include "mkldnn_graph_dumper.h"
 #include <legacy/cnn_network_impl.hpp>
 #include <legacy/ie_util_internal.hpp>
-#include <legacy/ie_ngraph_utils.hpp>
+#include <ie_ngraph_utils.hpp>
 #include "exec_graph_info.hpp"
 #include "mkldnn_debug.h"
 #include "generic_ie.hpp"
