+++ inference-engine/src/multi_device/CMakeLists.txt
@@ -12,7 +12,7 @@ ie_add_plugin(NAME ${TARGET_NAME}
               SOURCES ${SOURCES} ${HEADERS}
               VERSION_DEFINES_FOR multi_device.cpp)
 
-target_link_libraries(${TARGET_NAME} PRIVATE inference_engine)
+target_link_libraries(${TARGET_NAME} PRIVATE inference_engine_legacy inference_engine)
 
 set_ie_threading_interface_for(${TARGET_NAME})
