+++ inference-engine/src/plugin_api/generic_ie.hpp
@@ -21,7 +21,7 @@ namespace InferenceEngine {
 class IShapeInferExtension;
 using IShapeInferExtensionPtr = std::shared_ptr<IShapeInferExtension>;
 
-}
+}  // namespace InferenceEngine
 
 namespace ngraph {
 namespace op {
@@ -133,4 +133,3 @@ private:
 
 }  // namespace op
 }  // namespace ngraph
-
