+++ inference-engine/src/readers/ir_reader/ie_ir_parser.cpp
@@ -27,7 +27,7 @@
 #include <cpp/ie_cnn_network.h>
 #include "ie_blob_stream.hpp"
 #include "caseless.hpp"
-#include <legacy/ie_ngraph_utils.hpp>
+#include <ie_ngraph_utils.hpp>
 #include "generic_ie.hpp"
 #include "precision_utils.h"
 #include "blob_factory.hpp"
