+++ inference-engine/src/readers/ir_reader_v7/CMakeLists.txt
@@ -29,7 +29,7 @@ target_include_directories(${TARGET_NAME} PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/"
                                                   "${IE_MAIN_SOURCE_DIR}/src/readers/ir_reader")
 
 target_link_libraries(${TARGET_NAME} PRIVATE inference_engine_reader_api inference_engine_plugin_api
-                                             inference_engine pugixml openvino::itt)
+                                             inference_engine_legacy inference_engine pugixml openvino::itt)
 
 ie_add_api_validator_post_build_step(TARGET ${TARGET_NAME})
