+++ inference-engine/src/readers/ir_reader_v7/ie_cnn_net_reader_impl.cpp
@@ -3,6 +3,7 @@
 //
 
 #include <file_utils.h>
+#include <description_buffer.hpp>
 #include <ie_cnn_net_reader_impl.h>
 #include <ie_blob_stream.hpp>
